Lazy load `@lumino/datagrid` (#14037)
* load datagrid lazily in csv

* handle rapid disposal in csv tests

* linting

* address review comments
