+++ dev_mode/webpack.config.js
@@ -292,7 +292,7 @@ module.exports = [
         chunks: 'all',
         cacheGroups: {
           jlab_core: {
-            test: /[\\/]node_modules[\\/]@(jupyterlab|lumino)[\\/]/,
+            test: /[\\/]node_modules[\\/]@(jupyterlab|lumino(?!\/datagrid))[\\/]/,
             name: 'jlab_core'
           }
         }
