+++ packages/csvviewer-extension/src/index.ts
@@ -18,20 +18,19 @@ import {
   WidgetTracker
 } from '@jupyterlab/apputils';
 import {
-  CSVDelimiter,
-  CSVViewer,
   CSVViewerFactory,
-  TextRenderConfig,
   TSVViewerFactory
-} from '@jupyterlab/csvviewer';
+} from '@jupyterlab/csvviewer/lib/widget';
+import { CSVDelimiter } from '@jupyterlab/csvviewer/lib/toolbar';
+import type { CSVViewer } from '@jupyterlab/csvviewer';
+import type { TextRenderConfig } from '@jupyterlab/csvviewer';
 import { DocumentRegistry, IDocumentWidget } from '@jupyterlab/docregistry';
 import { ISearchProviderRegistry } from '@jupyterlab/documentsearch';
 import { IMainMenu } from '@jupyterlab/mainmenu';
 import { IObservableList } from '@jupyterlab/observables';
 import { ISettingRegistry } from '@jupyterlab/settingregistry';
 import { ITranslator } from '@jupyterlab/translation';
-import { DataGrid } from '@lumino/datagrid';
-import { CSVSearchProvider } from './searchprovider';
+import type { DataGrid } from '@lumino/datagrid';
 
 /**
  * The name of the factories that creates widgets.
@@ -156,7 +155,16 @@ function activateCsv(
 
   app.docRegistry.addWidgetFactory(factory);
   const ft = app.docRegistry.getFileType('csv');
-  factory.widgetCreated.connect((sender, widget) => {
+
+  let searchProviderInitialized = false;
+
+  factory.widgetCreated.connect(async (sender, widget) => {
+    if (searchRegistry && !searchProviderInitialized) {
+      const { CSVSearchProvider } = await import('./searchprovider');
+      searchRegistry.add('csv', CSVSearchProvider);
+      searchProviderInitialized = true;
+    }
+
     // Track the widget.
     void tracker.add(widget);
     // Notify the widget tracker if restore data needs to update.
@@ -222,9 +230,6 @@ function activateCsv(
       id: CommandIDs.CSVGoToLine,
       isEnabled
     });
-  }
-  if (searchRegistry) {
-    searchRegistry.add('csv', CSVSearchProvider);
   }
 }
 
@@ -300,7 +305,16 @@ function activateTsv(
 
   app.docRegistry.addWidgetFactory(factory);
   const ft = app.docRegistry.getFileType('tsv');
-  factory.widgetCreated.connect((sender, widget) => {
+
+  let searchProviderInitialized = false;
+
+  factory.widgetCreated.connect(async (sender, widget) => {
+    if (searchRegistry && !searchProviderInitialized) {
+      const { CSVSearchProvider } = await import('./searchprovider');
+      searchRegistry.add('tsv', CSVSearchProvider);
+      searchProviderInitialized = true;
+    }
+
     // Track the widget.
     void tracker.add(widget);
     // Notify the widget tracker if restore data needs to update.
@@ -367,9 +381,6 @@ function activateTsv(
       isEnabled
     });
   }
-  if (searchRegistry) {
-    searchRegistry.add('tsv', CSVSearchProvider);
-  }
 }
 
 /**
@@ -386,7 +397,6 @@ namespace Private {
    * The light theme for the data grid.
    */
   export const LIGHT_STYLE: DataGrid.Style = {
-    ...DataGrid.defaultStyle,
     voidColor: '#F3F3F3',
     backgroundColor: 'white',
     headerBackgroundColor: '#EEEEEE',
@@ -399,7 +409,6 @@ namespace Private {
    * The dark theme for the data grid.
    */
   export const DARK_STYLE: DataGrid.Style = {
-    ...DataGrid.defaultStyle,
     voidColor: 'black',
     backgroundColor: '#111111',
     headerBackgroundColor: '#424242',
