+++ packages/csvviewer/src/widget.ts
@@ -9,18 +9,11 @@ import {
   IDocumentWidget
 } from '@jupyterlab/docregistry';
 import { PromiseDelegate } from '@lumino/coreutils';
-import {
-  BasicKeyHandler,
-  BasicMouseHandler,
-  BasicSelectionModel,
-  CellRenderer,
-  DataGrid,
-  TextRenderer
-} from '@lumino/datagrid';
+import type * as DataGridModule from '@lumino/datagrid';
 import { Message } from '@lumino/messaging';
 import { ISignal, Signal } from '@lumino/signaling';
 import { PanelLayout, Widget } from '@lumino/widgets';
-import { DSVModel } from './model';
+import type * as DSVModelModule from './model';
 import { CSVDelimiter } from './toolbar';
 
 /**
@@ -57,7 +50,7 @@ export class TextRenderConfig {
   /**
    * horizontalAlignment of the text
    */
-  horizontalAlignment: TextRenderer.HorizontalAlignment;
+  horizontalAlignment: DataGridModule.TextRenderer.HorizontalAlignment;
 }
 
 /**
@@ -67,7 +60,7 @@ export class TextRenderConfig {
  * to set the background color of cells matching the search text.
  */
 export class GridSearchService {
-  constructor(grid: DataGrid) {
+  constructor(grid: DataGridModule.DataGrid) {
     this._grid = grid;
     this._query = null;
     this._row = 0;
@@ -88,7 +81,7 @@ export class GridSearchService {
    */
   cellBackgroundColorRendererFunc(
     config: TextRenderConfig
-  ): CellRenderer.ConfigFunc<string> {
+  ): DataGridModule.CellRenderer.ConfigFunc<string> {
     return ({ value, row, column }) => {
       if (this._query) {
         if ((value as string).match(this._query)) {
@@ -216,7 +209,7 @@ export class GridSearchService {
     return this._query;
   }
 
-  private _grid: DataGrid;
+  private _grid: DataGridModule.DataGrid;
   private _query: RegExp | null;
   private _row: number;
   private _column: number;
@@ -234,11 +227,22 @@ export class CSVViewer extends Widget {
   constructor(options: CSVViewer.IOptions) {
     super();
 
-    const context = (this._context = options.context);
-    const layout = (this.layout = new PanelLayout());
+    this._context = options.context;
+    this.layout = new PanelLayout();
 
     this.addClass(CSV_CLASS);
 
+    void this.initialize();
+  }
+
+  protected async initialize(): Promise<void> {
+    const layout = this.layout as PanelLayout;
+    if (this.isDisposed || !layout) {
+      return;
+    }
+    const { BasicKeyHandler, BasicMouseHandler, DataGrid } =
+      await Private.ensureDataGrid();
+    this._defaultStyle = DataGrid.defaultStyle;
     this._grid = new DataGrid({
       defaultSizes: {
         rowHeight: 24,
@@ -263,16 +267,15 @@ export class CSVViewer extends Widget {
     this._searchService = new GridSearchService(this._grid);
     this._searchService.changed.connect(this._updateRenderer, this);
 
-    void this._context.ready.then(() => {
-      this._updateGrid();
-      this._revealed.resolve(undefined);
-      // Throttle the rendering rate of the widget.
-      this._monitor = new ActivityMonitor({
-        signal: context.model.contentChanged,
-        timeout: RENDER_TIMEOUT
-      });
-      this._monitor.activityStopped.connect(this._updateGrid, this);
+    await this._context.ready;
+    await this._updateGrid();
+    this._revealed.resolve(undefined);
+    // Throttle the rendering rate of the widget.
+    this._monitor = new ActivityMonitor({
+      signal: this._context.model.contentChanged,
+      timeout: RENDER_TIMEOUT
     });
+    this._monitor.activityStopped.connect(this._updateGrid, this);
   }
 
   /**
@@ -300,17 +303,17 @@ export class CSVViewer extends Widget {
       return;
     }
     this._delimiter = value;
-    this._updateGrid();
+    void this._updateGrid();
   }
 
   /**
    * The style used by the data grid.
    */
-  get style(): DataGrid.Style {
+  get style(): DataGridModule.DataGrid.Style {
     return this._grid.style;
   }
-  set style(value: DataGrid.Style) {
-    this._grid.style = value;
+  set style(value: DataGridModule.DataGrid.Style) {
+    this._grid.style = { ...this._defaultStyle, ...value };
   }
 
   /**
@@ -318,7 +321,7 @@ export class CSVViewer extends Widget {
    */
   set rendererConfig(rendererConfig: TextRenderConfig) {
     this._baseRenderer = rendererConfig;
-    this._updateRenderer();
+    void this._updateRenderer();
   }
 
   /**
@@ -356,10 +359,12 @@ export class CSVViewer extends Widget {
   /**
    * Create the model for the grid.
    */
-  private _updateGrid(): void {
+  private async _updateGrid(): Promise<void> {
+    const { BasicSelectionModel } = await Private.ensureDataGrid();
+    const { DSVModel } = await Private.ensureDSVModel();
     const data: string = this._context.model.toString();
     const delimiter = this._delimiter;
-    const oldModel = this._grid.dataModel as DSVModel;
+    const oldModel = this._grid.dataModel as DSVModelModule.DSVModel;
     const dataModel = (this._grid.dataModel = new DSVModel({
       data,
       delimiter
@@ -373,10 +378,11 @@ export class CSVViewer extends Widget {
   /**
    * Update the renderer for the grid.
    */
-  private _updateRenderer(): void {
+  private async _updateRenderer(): Promise<void> {
     if (this._baseRenderer === null) {
       return;
     }
+    const { TextRenderer } = await Private.ensureDataGrid();
     const rendererConfig = this._baseRenderer;
     const renderer = new TextRenderer({
       textColor: rendererConfig.textColor,
@@ -393,7 +399,8 @@ export class CSVViewer extends Widget {
   }
 
   private _context: DocumentRegistry.Context;
-  private _grid: DataGrid;
+  private _grid: DataGridModule.DataGrid;
+  private _defaultStyle: typeof DataGridModule.DataGrid.defaultStyle;
   private _searchService: GridSearchService;
   private _monitor: ActivityMonitor<DocumentRegistry.IModel, void> | null =
     null;
@@ -472,14 +479,6 @@ export namespace CSVDocumentWidget {
   }
 }
 
-namespace Private {
-  export function createContent(
-    context: DocumentRegistry.IContext<DocumentRegistry.IModel>
-  ): CSVViewer {
-    return new CSVViewer({ context });
-  }
-}
-
 /**
  * A widget factory for CSV widgets.
  */
@@ -530,5 +529,35 @@ export class TSVViewerFactory extends CSVViewerFactory {
       delimiter,
       translator: this.translator
     });
+  }
+}
+
+namespace Private {
+  let gridLoaded: PromiseDelegate<typeof DataGridModule> | null = null;
+  let modelLoaded: PromiseDelegate<typeof DSVModelModule> | null = null;
+
+  /**
+   * Lazily load the datagrid module when the first grid is requested.
+   */
+  export async function ensureDataGrid(): Promise<typeof DataGridModule> {
+    if (gridLoaded == null) {
+      gridLoaded = new PromiseDelegate();
+      gridLoaded.resolve(await import('@lumino/datagrid'));
+    }
+    return gridLoaded.promise;
+  }
+
+  export async function ensureDSVModel(): Promise<typeof DSVModelModule> {
+    if (modelLoaded == null) {
+      modelLoaded = new PromiseDelegate();
+      modelLoaded.resolve(await import('./model'));
+    }
+    return modelLoaded.promise;
+  }
+
+  export function createContent(
+    context: DocumentRegistry.IContext<DocumentRegistry.IModel>
+  ): CSVViewer {
+    return new CSVViewer({ context });
   }
 }
